init Ass7.1

+++ src/src/Controller/HotelController.php
@@ -23,6 +23,7 @@ class HotelController extends AbstractController {
 
     #[Route('/new', name: 'app_hotel_new', methods: ['GET', 'POST'])]
     public function new(Request $request, HotelRepository $hotelRepository): Response {
+        $this->denyAccessUnlessGranted("ROLE_HOTEL_OWNER");
         $hotel = new Hotel();
         $form = $this->createForm(HotelType::class, $hotel);
         $form->handleRequest($request);
@@ -59,6 +60,8 @@ class HotelController extends AbstractController {
 
     #[Route('/{id}/edit', name: 'app_hotel_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Hotel $hotel, HotelRepository $hotelRepository): Response {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR","ROLE_HOTEL_OWNER");
+        $this->denyAccessUnlessGranted('edit',$hotel);
         $form = $this->createForm(HotelType::class, $hotel);
         $form->handleRequest($request);
 
@@ -77,6 +80,8 @@ class HotelController extends AbstractController {
 
     #[Route('/{id}', name: 'app_hotel_delete', methods: ['POST'])]
     public function delete(Request $request, Hotel $hotel, HotelRepository $hotelRepository): Response {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR","ROLE_HOTEL_OWNER");
+        $this->denyAccessUnlessGranted('delete',$hotel);
         if ($this->isCsrfTokenValid('delete'.$hotel->getId(), $request->request->get('_token'))) {
             $hotelRepository->remove($hotel);
         }
